// sanity/types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;
// ...
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type HomePageQueryResult = {
  hero: {
    heading: string | null;
    image: {
      asset: SanityImageAssetReference | null;
      hotspot: SanityImageHotspot | null;
      crop: SanityImageCrop | null;
      alt: string | null;
    } | null;
  } | null;
  about: {
    heading: string | null;
    body: Array<{
      children?: Array<{
        marks?: Array<string>;
        text?: string;
        _type: "span";
        _key: string;
      }>;
      style?: "normal";
      listItem?: never;
      markDefs?: Array<{
        href?: string;
        _type: "link";
        _key: string;
      }>;
      level?: number;
      _type: "block";
      _key: string;
    }> | null;
    images: Array<{
      _key: string;
      asset: SanityImageAssetReference | null;
      hotspot: SanityImageHotspot | null;
      crop: SanityImageCrop | null;
      alt: string | null;
    }> | null;
  } | null;
  faq: {
    heading: string | null;
    items: Array<{
      _key: string;
      question: string | null;
      answer: string | null;
    }> | null;
  } | null;
  testimonials: {
    heading: string | null;
    items: Array<{
      _key: string;
      quote: string | null;
      name: string | null;
      role: string | null;
      photo: {
        asset: SanityImageAssetReference | null;
        hotspot: SanityImageHotspot | null;
        crop: SanityImageCrop | null;
        alt: string | null;
      } | null;
    }> | null;
  } | null;
} | null;